pyproject.toml
[tool.poetry]
name = "repl_python3_tripple-fard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.2.4"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
jishaku = "^2.3.0"
utils = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
